pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "htbctl"
version = "0.1.0"
description = "HackTheBox machine lifecycle automation: spawn, attack, stop."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "requests>=2.28",
    "python-dotenv>=1.0",
]
authors = [{ name = "drewrukin" }]
keywords = ["hackthebox", "htb", "ctf", "pentest", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]

[project.optional-dependencies]
dev = ["pytest"]

[project.scripts]
htbctl = "htbctl.cli:main"

[project.urls]
Homepage = "https://github.com/drewrukin/htbctl"
Issues = "https://github.com/drewrukin/htbctl/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
